fixed error in ecs serializer

+++ dependencies/sf2/include/sf2/formats/json_writer.hpp
@@ -217,11 +217,11 @@ namespace format {
 	}
 
 	inline void Json_writer::write(uint8_t v) {
-		_write(v);
+		_write((uint16_t) v);
 	}
 
 	inline void Json_writer::write(int8_t v) {
-		_write(v);
+		_write((int16_t) v);
 	}
 
 	inline void Json_writer::write(uint16_t v) {
